io: implement io_open, io_close and io_readlines
Rework the implementation of line reading and provide a new io_readlines
routing which calls a callback argument for each read line.

Abstract its usage inside a new click_read function for click parsing.

For block updates, no matter if the block is persistent or not, use the
io_readlines function to read an arbitrary count of lines, and pass the
callback corresponding to the block format.

Add io_open and io_close routines to abstract these syscalls and their
error checking. io_open only opens read-only and non-blocking.