block: provide click helper
Add a block_touch helper to reset the timestamp outside the spawn
function and a block_click helper called once a block is clicked.

Now that we have dynamic properties, we can inject a click's "x", "y",
and "button" properties into the block, and thus get rid of the click
structure from within the spawn code.